requirements.txt
pydantic>=2.6.0
pyyaml>=6.0.1
requests>=2.31.0
loguru>=0.7.2
pycryptodome>=3.20.0
cachetools>=5.3.3
xmltodict>=0.13.0
